Add ManifestGenerator: Cecil-free manifest generation (#10807)
Split into focused classes in Microsoft.Android.Sdk.TrimmableTypeMap:

- ManifestGenerator: orchestration (load template, call builders, write output)
- ManifestConstants: shared AndroidNs and AttName
- PropertyMapper: data-driven property mapping (7 arrays, MappingKind enum)
- AndroidEnumConverter: 9 enum-to-string converters
- ComponentElementBuilder: Activity/Service/Receiver/Provider/Instrumentation XML
- AssemblyLevelElementBuilder: permissions, uses-permissions, features, libraries

Features: MainLauncher, runtime provider (with dedup), template merging,
ManifestPlaceholders, debuggable/extractNativeLibs, XA4213 validation.

30 unit tests (xunit, 130ms).
