+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Generator/ManifestConstants.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+using System.Xml.Linq;
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap;
+
+static class ManifestConstants
+{
+	public static readonly XNamespace AndroidNs = "http://schemas.android.com/apk/res/android";
+	public static readonly XName AttName = AndroidNs + "name";
+}
